juego 100%funcional

+++ TFBProyecto/app/src/main/java/com/example/tfb/Usuario.kt
@@ -6,15 +6,15 @@ package com.example.tfb
 import android.provider.ContactsContract.CommonDataKinds.Photo
 import com.example.tfb.Enumerados.ProviderType
 
-data class Usuario(val nombre: String, val email: String, val provider: ProviderType, val maxscore: Double) {
+data class Usuario(val nombre: String, val email: String, val provider: ProviderType, val maxscore: Int) {
 
     companion object {
         // Static variables or functions can be defined here
         var currentUsuario: Usuario? = null
         fun crearUsuarioInvitado(): Usuario {
-            return Usuario("Invitado", "", ProviderType.NINGUNO, 0.0)
+            return Usuario("Invitado", "", ProviderType.NINGUNO, 0)
         }
-        fun crearUsuario(nombre: String, email: String, provider: ProviderType, maxscore: Double): Usuario {
+        fun crearUsuario(nombre: String, email: String, provider: ProviderType, maxscore: Int): Usuario {
             return Usuario(nombre, email, provider, maxscore)
         }
